Add TensorBoard logs to 01_adding_task.py and 02_frequency_discrimination_task.py
