+++ src/01_adding_task.py
@@ -9,6 +9,9 @@ Validation is performed on data generated on the fly.
 from __future__ import absolute_import
 from __future__ import print_function
 
+import os
+import datetime
+
 import random
 import numpy as np
 
@@ -124,6 +127,9 @@ def train():
 
     sess = tf.Session()
 
+    log_dir = os.path.join(FLAGS.logdir, datetime.datetime.now().strftime("%Y%m%d-%H%M%S"))
+    valid_writer = tf.summary.FileWriter(log_dir + '/val')
+
     sess.run(tf.global_variables_initializer())
 
     try:
@@ -155,6 +161,11 @@ def train():
                         valid_steps += FLAGS.sequence_length
                 valid_error /= FLAGS.validation_batches
                 valid_steps /= FLAGS.validation_batches
+
+                valid_writer.add_summary(scalar_summary('error', valid_error), num_iters)
+                valid_writer.add_summary(scalar_summary('used_samples', valid_steps / FLAGS.sequence_length), num_iters)
+                valid_writer.flush()
+
                 print("Iteration %d, "
                       "validation error: %.7f, "
                       "validation samples: %.2f%%" % (num_iters,
