+++ src/02_frequency_discrimination_task.py
@@ -8,6 +8,9 @@ Batches are stratified. Validation is performed on data generated on the fly.
 from __future__ import absolute_import
 from __future__ import print_function
 
+import os
+import datetime
+
 import random
 import numpy as np
 
@@ -130,6 +133,9 @@ def train():
 
     sess = tf.Session()
 
+    log_dir = os.path.join(FLAGS.logdir, datetime.datetime.now().strftime("%Y%m%d-%H%M%S"))
+    valid_writer = tf.summary.FileWriter(log_dir + '/val')
+
     sess.run(tf.global_variables_initializer())
 
     try:
@@ -165,6 +171,11 @@ def train():
                         valid_steps += SEQUENCE_LENGTH
                 valid_accuracy /= FLAGS.validation_batches
                 valid_steps /= FLAGS.validation_batches
+
+                valid_writer.add_summary(scalar_summary('accuracy', valid_accuracy), num_iters)
+                valid_writer.add_summary(scalar_summary('used_samples', valid_steps / SEQUENCE_LENGTH), num_iters)
+                valid_writer.flush()
+
                 print("Iteration %d, "
                       "validation accuracy: %.2f%%, "
                       "validation samples: %.2f (%.2f%%)" % (num_iters,
